fix(docsite): make canary target statically exportable
The one-deploy build (build-versioned.mjs) exports canary as a static site
(output:'export'), but several routes used features that force dynamic
rendering, so the export failed. Verified by running the canary export to
completion: it now prerenders all 268 pages and emits out/ with basePath
/canary baked into asset URLs.

Fixes, each keeping the LATEST server build identical:
- robots.ts / sitemap.ts: add `export const dynamic = 'force-static'`. Both are
  already build-time static (derived from generated registries + SITE_URL);
  the export just needs them declared static to emit robots.txt / sitemap.xml.
- themes/page.tsx: read `?theme=` on the client instead of server
  `await searchParams` (which forces dynamic rendering). Split into a server
  page (keeps metadata) + a client ThemesExplorer that resolves the deep-link
  seed via useSearchParams, wrapped in Suspense. Deep links still preselect the
  picker; the read just happens in the browser.
- templates/page.tsx: wrap the useSearchParams client body in a Suspense
  boundary (required by the static export; the server build tolerated its
  absence).

Other useSearchParams consumers already sit under a Suspense ancestor
(PlaygroundClient, ComponentDetailClient, ThemesExplorer) so no change needed.
